+++ psumsim/rand.py
@@ -30,9 +30,10 @@ class sinusoidal_gen(sp.stats.rv_continuous):
 		*scale*. This is the least we need to define.
 		
 		The *CDF* describes: with which probability do I draw a number smaller
-		than *x* from the random process. For :math:`x < -1`, that is *0*. A
-		sinusoid never yields values smaller than *-1*. For :math:`x > 1`, that
-		is *1*, as all values of a sinusoid are below or eqal *1*.
+		than or equal to *x* from the random process. For :math:`x < -1`, that
+		is *0*. A sinusoid never yields values smaller than *-1*. For
+		:math:`x > 1`, that is *1*, as all values of a sinusoid are below or
+		eqal *1*.
 		
 		For values in-between, things are tricky. We have to look at regions
 		where :math:`sin(k) < x` and check which length compared to
@@ -174,3 +175,145 @@ sinusoidal = sinusoidal_gen(name="sinusoidal")
 
 `scipy.stats` does this as well. Use this as the default generator and call
 this instance (not the class) to create frozen distributions."""
+
+
+class fullscale_gen(sp.stats.rv_continuous):
+	r"""Define random process of variable having a constant full scale value.
+	
+	Similar to `sinusoidal_gen`.
+	
+	Use *loc* to define the full-scale value.
+	"""
+	
+	SAMPLE_DTYPE = "float"
+	"""`str` : Datatype accepeted by `numpy` functions. Used in `_rvs`. Use
+	platform's best floating point."""
+	
+	def _cdf(self, x):
+		r"""Cumulative density function.
+		
+		Similar to `sinusoidal_gen._cdf`.
+		
+		We here only need to implement a step function with step at location
+		zero. *loc* was already applied for us when this gets called. The *CDF*
+		gives the probability to yield a value smaller *or equal* to *x*. So
+		if *x* is *loc*, we already return a *1* probability.
+
+		Parameters
+		----------
+		x : `numpy.ndarray`
+			Values to compute *CDF* for.
+
+		Returns
+		-------
+		ret : `numpy.ndarray`
+			Same shape and dtype as *x*, but with computed *CDF* values.
+
+		"""
+		
+		#Implement step function
+		ret = np.where(x >= 0, 1, 0)
+		#Cast dtype. If it is already the one of input value, keep without copy.
+		ret = ret.astype(dtype=x.dtype, copy=False)
+		
+		return ret
+		
+	def _pdf(self, x):
+		r"""Probability density function.
+		
+		Similar to `sinusoidal_gen._pdf`.
+		
+		Derivative of `_cdf`. Ideally, this would be a Dirac pulse, as that
+		also has the correct area of *1*. But we here can only output *0*
+		all the time and *Inf* at *0*.
+		
+		Parameters
+		----------
+		x : `numpy.ndarray`
+			Values to compute *PDF* for.
+
+		Returns
+		-------
+		ret : `numpy.ndarray`
+			Same shape and dtype as *x*, but with computed *PDF* values.
+
+		"""
+		
+		#Output with correct shape and dtype, but zero everywhere
+		ret = np.zeros_like(x)
+		#Where x was zero, the derivative is inf
+		np.putmask(ret, mask=(x == 0), values=np.inf)
+		
+		return ret
+
+	def _ppf(self, x):
+		r"""Percent point function.
+		
+		Similar to `sinusoidal_gen._pdf`.
+		
+		This is not implemented. Should be the inverse of `_cdf`, but the *CDF*
+		yields *0* for all :math:`x < loc` and *1* for all :math:`x \ge loc`.
+		Which value to return then? And for all other values, the ivnerse would
+		be *NaN*.
+
+		Parameters
+		----------
+		x : `numpy.ndarray`
+			Values to compute *PPF* for.
+
+		Returns
+		-------
+		ret : `None`
+			
+		Raises
+		------
+		`NotImplementedError`
+			Always.
+
+		"""
+		
+		raise NotImplementedError(
+				"The PPF for fullscale random process is not defined."
+		)
+
+	def _rvs(self, size=None, random_state=None):
+		"""Define how random numbers are drawn from this process.
+		
+		Overwrites `scipy.stats.rv_continuous`.*_rvs* and is called by
+		`scipy.stats.rv_continuous.rvs`.
+		
+		`_ppf` cannot be called, so the default definiton of this method does
+		not work, so we create our own one.
+		
+		Functionality is easy: this random process draws the same number all
+		the time. And *loc* is even applied for us. So we just return zeros
+		all the time.
+		
+
+		Parameters
+		----------
+		size : (`tuple` of `int`), `None`, optional
+			Size of array to generate. `None` generates a 0D array.
+			The default is `None`.
+		random_state : `object`, optional
+			The random number generator. Not used and only added for
+			interface compatibility with `scipy.stats.rv_continuous`.*_rvs*.
+			The default is `None`.
+
+		Returns
+		-------
+		ret : `numpy.ndarray`
+			One array with dtype `SAMPLE_DTYPE` and *0* values, which will be
+			moved to *loc*. Shape set by *size*.
+
+		"""
+		
+		ret = np.zeros(shape=size, dtype=self.SAMPLE_DTYPE)
+
+		return ret
+
+
+fullscale = fullscale_gen(name="fullscale")
+"""`fullscale_gen` : Default instance of number generator.
+
+Like `sinusoidal`."""